requirements.txt
chromadb==0.5.0
sentence-transformers==2.7.0
openai==1.30.1
python-dotenv==1.0.1
pymupdf==1.24.3
tqdm==4.66.4
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2